fix: move movie service directory; add annotation
