+++ src/main/java/com/jp/movieapi/component/movie/service/MovieServiceImpl.java
@@ -1,10 +1,13 @@
-package com.jp.movieapi.component.movie.service.movie;
+package com.jp.movieapi.component.movie.service;
 
 import java.util.List;
+
+import org.springframework.stereotype.Service;
 
 import com.jp.movieapi.component.movie.domain.Movie;
 import com.jp.movieapi.component.movie.domain.MovieRepository;
 
+@Service
 public class MovieServiceImpl implements MovieService {
     public final MovieRepository movieRepository;
