+++ src/main/java/com/jp/movieapi/internal/api/controllers/movie/MovieController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.jp.movieapi.component.movie.domain.Movie;
-import com.jp.movieapi.component.movie.service.movie.MovieService;
+import com.jp.movieapi.component.movie.service.MovieService;
 
 @RestController
 @RequestMapping("/api/movies")
